+++ docs/.vitepress/theme/components/bootstrap-vue/table-child/BTableHeader.vue
@@ -1,0 +1,33 @@
+<template>
+    <thead>
+        <tr>
+            <th v-for="header in headers" :key="header.key" @click="handleSort(header)">
+                <span>{{ header.label }}</span>
+                <BTableSortIcon 
+                v-if="!header.unsortable"
+                :sort-key="sortKey"
+                :sort-order="sortOrder"
+                :header-key="header.key"
+                />
+            </th>
+        </tr>
+    </thead>
+</template>
+
+<script setup>
+import BTableSortIcon from './BTableSortIcon.vue'
+
+const props = defineProps({
+    headers: Array,
+    sortKey: String,
+    sortOrder: String
+})
+
+const emit = defineEmits(['sort'])
+
+function handleSort(header) {
+    if (!header.unsortable) {
+        emit('sort', header)
+    }
+}
+</script>
